Make runtest wait for the server to start
Instead of having a fixed startup time of 1 second, we can use curl to
check if the server is responding to requests. We wait 0.1 seconds
between request attempts, and run the test suite as soon as we get a
successful response. If we don't get a response after 5 seconds, then
we give up.